refactor: move files according to their specifity
