+++ src/common/endpoint.constants.js
