+++ src/common/endpoint.factory.js
